Fix pace precision loss for imperial unit splits
Changed pace storage from integer to double precision to eliminate
rounding errors when converting between metric and imperial units.

Previously, storing pace as integer truncated fractional seconds during
calculation (e.g., 435.545 → 435 sec/km), which compounded during
metric-to-imperial conversion. This caused displayed pace to mismatch
actual split times by 1-3 seconds for imperial users (e.g., a 1.00 mi
split in 11:41 showed pace as 11:40 /mi).

Backend changes:
- Changed WorkoutSplit.PaceS and Workout.AvgPaceS from int to double
- Removed integer casts from pace calculations in parser services
- Created migration to alter database column types
- Updated DTOs and test data to use double precision

Frontend changes:
- Updated formatPace() to round total seconds before formatting
- Prevents floating-point precision issues in display

Impact: Pace values now accurately reflect split times for both
metric and imperial units, with no precision loss from truncation.

+++ api/Endpoints/WorkoutsEndpoints.cs
@@ -28,10 +28,10 @@ public static class WorkoutsEndpoints
         public DateTime StartedAt { get; set; }
         public int DurationS { get; set; }
         public double DistanceM { get; set; }
-        public int AvgPaceS { get; set; }
+        public double AvgPaceS { get; set; }
         public double? SimilarityScore { get; set; }
         public int? TimeDifferenceS { get; set; } // Negative = faster, Positive = slower
-        public int? PaceDifferenceS { get; set; } // Negative = faster pace
+        public double? PaceDifferenceS { get; set; } // Negative = faster pace
         public int? RelativeEffort { get; set; }
         public double? ElevGainM { get; set; }
     }
